moved repeater up one level.
